backend/requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
python-dotenv
python-jose
sqlalchemy
psycopg2
